SLA-109 Use special constant for operator account inside the action list (#496)
* SLA-109 Use a special constant for the operator account inside the action list

NeonEVM puts `Action::ExternalInstriction` with the current operator on each iteration. If different operators executed iterations it leads to a problem, because on the final iteration, we have an action with the previous operator.

Solution: we put a special constant to the action list and change this constant with the real operator in the final iteration.